read_bytes: some more size_t stuff

+++ io.c
@@ -44,20 +44,20 @@ uint8_t get_byte(FILE *fp)
 
 int read_bytes(FILE *fp, void *data, unsigned int sz)
 {
-  int status = fread(data, 1, sz, fp);
+  size_t status = fread(data, 1, (size_t) sz, fp);
 
-  if (status == EOF)
+  if ((int) status == EOF)
   {
     fprintf(stderr, "End of file.\n");
     exit(EXIT_FAILURE);
   }
-  else if (status != sz)
+  else if (status != (size_t) sz)
   {
     fprintf(stderr, "Error reading bytes\n");
     exit(EXIT_FAILURE);
   }
 
-  return status;
+  return (int) status;
 }
 
 /**
